Refactor volume expand
* Perform validations holding the lock
* Emit event on volume expand
* Move txn step functions to it's own file
* Notifying clients of volfile change must be done on all nodes
